pyproject.toml
[tool.poetry]
name = "django-signal-webhooks"
version = "0.2.4"
description = "Add webhooks to django using signals."
authors = [
    "Matti Lamppu <lamppu.matti.akseli@gmail.com>",
]
packages = [
    { include = "signal_webhooks" },
]
license = "MIT"
readme = "README.md"
homepage = "https://mrthearman.github.io/django-signal-webhooks/"
repository = "https://github.com/MrThearMan/django-signal-webhooks"
keywords = [
    "django",
    "signal",
    "webhooks"
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]
include = [
    "README.md",
    "LICENSE",
]
exclude = [
    "tests",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/MrThearMan/django-signal-webhooks/issues"

[tool.poetry.dependencies]
python = ">=3.9,<4"
Django = ">=3.2"
asgiref = ">=3.5.0"
httpx = ">=0.23.0"
cryptography = ">=36.0.0"
django-settings-holder = ">=0.1.0"
djangorestframework = { version = ">=3.13.0", optional = true}

[tool.poetry.group.test.dependencies]
pytest = "7.4.3"
coverage = "7.3.2"
freezegun = "^1.2.2"
pytest-django = "4.5.2"
pre-commit = "3.5.0"
tox = "4.11.3"
tox-gh-actions = "3.1.3"

[tool.poetry.group.docs.dependencies]
mkdocs = "1.5.3"
pymdown-extensions = "10.3.1"
mkdocs-mermaid2-plugin = "1.1.1"

[tool.poetry.group.lint.dependencies]
mypy = "1.6.1"
django-stubs = "4.2.6"
djangorestframework-stubs = "3.14.4"

[tool.poetry.extras]
drf = ["djangorestframework"]

[tool.ruff]
fix = true
line-length = 120
exclude = [
    "tests/*",
]
select = [
    "A",        # flake8-builtins
    "ANN",      # flake8-annotations
    "ARG",      # flake8-unused-arguments
    "B",        # flake8-bugbear
    "BLE",      # flake8-blind-except
    "C90",      # mccabe
    "C4",       # flake8-comprehensions
    "COM",      # flake8-commas
    "D200",     # pydocstyle: One-line docstring should fit on one line
    "D201",     # pydocstyle: No blank lines allowed before function docstring (found {num_lines})
    "D202",     # pydocstyle: No blank lines allowed after function docstring (found {num_lines})
    "D209",     # pydocstyle: Multi-line docstring closing quotes should be on a separate line
    "D213",     # pydocstyle: Multi-line docstring summary should start at the second line
    "DTZ",      # flake8-datetimez
    "E",        # pycodestyle errors
    "EM",       # flake8-errmsg
    "F",        # pyflakes
    "FBT",      # flake8-boolean-trap
    "I",        # isort
    "INP",      # flake8-no-pep420
    "ISC",      # flake8-implicit-str-concat
    "N",        # pep8-naming
    "PERF",     # perflint
    "PGH",      # pygrep-hooks
    "PIE",      # flake8-pie
    "PL",       # pylint
    "PT",       # flake8-pytest-style
    "PTH",      # flake8-use-pathlib
    "PYI",      # flake8-pyi
    "Q",        # flake8-quotes
    "RET",      # flake8-return
    "RSE",      # flake8-raise
    "RUF",      # ruff-specific rules
    "S",        # flake8-bandit
    "SIM",      # flake8-simplify
    "T20",      # flake8-print
    "TCH",      # flake8-type-checking
    "TRY",      # tryceratops
    "UP",       # pyupgrade
    "W",        # pycodestyle warnings
]
ignore = [
    "S101",     # assert is OK
    "S105",     # no hardcoded passwords
    "S311",     # random-module is OK.
    "ARG001",   # Anused function argument
    "ANN101",   # Missing type annotation for `self` in method
    "ANN102",   # Missing type annotation for `cls` in classmethod
    "ANN401",   # Any is allowed
    "RUF012",   # No ClassVar required
    #
    # Conflicting with ruff-format
    #
    "COM812",   # missing-trailing-comma
    "COM819",   # prohibited-trailing-comma
    "D206",     # indent-with-spaces
    "D300",     # triple-single-quotes
    "E111",     # indentation-with-invalid-multiple
    "E114",     # indentation-with-invalid-multiple-comment
    "E117",     # over-indented
    "ISC001",   # single-line-implicit-string-concatenation
    "ISC002",   # multi-line-implicit-string-concatenation
    "Q000",     # bad-quotes-inline-string
    "Q001",     # bad-quotes-multiline-string
    "Q002",     # bad-quotes-docstring
    "Q003",     # avoidable-escaped-quote
    "W191",     # tab-indentation
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"signal_webhooks/admin.py" = ["S105"]

[tool.ruff.isort]
known-third-party = [
    "django",
    "asgiref",
    "httpx",
    "cryptography",
    "django-settings-holder",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = "True"
warn_unused_configs = "True"
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[tool.coverage.run]
relative_files = true

[tool.coverage.report]
omit = [
    "tests/*",
    "docs/*",
    ".venv/*",
    ".tox/*",
]
exclude_lines = [
    "if TYPE_CHECKING:",
    "except ImportError:",
    "pragma: no cover",
]

[tool.pytest.ini_options]
addopts = "-vv -s --log-cli-level=INFO"
markers = [
    "e2e: Marks tests as e2e.",
]

[tool.django-stubs]
django_settings_module = "tests.project.settings"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{39, 310, 311, 312}-django{32, 40, 41, 42}
isolated_build = true

[gh-actions]
python =
    3.9: py39
    3.10: py310
    3.11: py311
    3.12: py312

[testenv]
allowlist_externals =
    poetry
setenv =
    PYTHONPATH = {toxinidir}
deps =
    django32: Django>=3.2,<4.0
    django40: Django>=4.0,<4.1
    django41: Django>=4.1,<4.2
    django42: Django>=4.2,<4.3

    asgiref>=3.5.0
    httpx>=0.23.0
    cryptography>=36.0.0
    django-settings-holder>=0.1.0
    djangorestframework>=3.13.0

    pytest
    coverage
    pytest-django
    freezegun
commands =
    coverage run -m pytest -m "not e2e"
"""

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
